simulator/requirements.txt
kafka-python==2.0.2
pydantic==2.6.1
python-dotenv==1.0.1
